+++ gnfinder/cmd/root.go
@@ -55,7 +55,7 @@ var rootCmd = &cobra.Command{
 		}
 
 		if len(args) == 0 {
-			cmd.Help()
+			_ = cmd.Help()
 			os.Exit(0)
 		}
 	},