+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add [#45]: return desired number of words before and after a name-candidate.
+
 ## [v0.9.1]
 
 - Add [#39]: Export to C shared library.
@@ -146,7 +148,10 @@ This document follows [changelog guidelines]
 [v0.5.1]: https://github.com/gnames/gnfinder/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/gnames/gnfinder/tree/v0.5.0
 
-[#42]: https://github.com/gnames/gnfinder/issues/42
+[#46]: https://github.com/gnames/gnfinder/issues/46
+[#45]: https://github.com/gnames/gnfinder/issues/45
+[#44]: https://github.com/gnames/gnfinder/issues/44
+[#43]: https://github.com/gnames/gnfinder/issues/43
 [#41]: https://github.com/gnames/gnfinder/issues/41
 [#40]: https://github.com/gnames/gnfinder/issues/40
 [#39]: https://github.com/gnames/gnfinder/issues/39